commit 18: port Setup affordances into unified Checklist
Second of three commits that fold GameSetupPanel + ChecklistGrid
into one tabbed surface. This commit teaches Checklist.tsx to render
the full Setup-mode UI when state.uiMode === "setup", extracts the
deck-swap row into its own component, and puts GameSetupPanel to
sleep.

Checklist.tsx — Setup-mode additions, all gated on inSetup:
- Player column headers become PlayerNameInput — editable with
  duplicate-name guard + per-column remove-×. Trailing "+" column
  at the end of each row adds players (empty <td> on card rows and
  on the case-file spot so the column count lines up).
- Hand-size row mirrors the former GameSetupPanel layout: one
  <input type="number"> per player with placeholder = defaultHandSizes,
  warning banner above the table when the column total diverges from
  the dealt-card count.
- Category header row gets InlineTextEdit + remove-× (disabled when
  only one category remains).
- Card name cell gets InlineTextEdit + remove-× (disabled on the
  last card in a category).
- "+ add card" row at the end of each category group, "+ add category"
  row at the end of the table body.
- colSpan on header / add-* rows is cardSpan = 1 + owners + (inSetup
  ? 1 : 0) so the trailing "+ add player" column doesn't make the
  rows ragged.

The cell grid itself (Y / N / blank, tooltips, cross-highlighting,
known-card toggle) is identical to Play mode — it's the same
table, more chrome.

PlayerNameInput and InlineTextEdit are duplicated from
GameSetupPanel for now. Commit 19 deletes the other copy when the
old panel goes away.

New src/ui/components/CardPackRow.tsx:
- Self-contained deck-swap row. Renders only when state.uiMode ===
  "setup" (returns null otherwise). Ported straight from the
  preset-row block in GameSetupPanel.

src/ui/Clue.tsx:
- <CardPackRow /> mounts between <TabBar /> and <Checklist />.
- Old GameSetupPanel + ChecklistGrid pair still mounted below as a
  documented safety net. GameSetupPanel now returns null, so it's
  effectively invisible; ChecklistGrid still shows the old
  deduction view. Both deleted in commit 19.

src/ui/components/GameSetupPanel.tsx:
- Body replaced with `return null`. File kept so commit 19's delete
  is a focused diff.

Verify: pnpm typecheck / test / lint / knip clean. Browser snapshot
in Setup mode: 2 role=tab buttons (Setup selected), 2 card-pack
choice buttons + "+ Save as card pack", 4 player-name <input>s in
the header, 4 hand-size <input type=number>s below, 3 "+ add card"
buttons (one per Classic category) + 1 "+ add category" button
below the body.
