+++ src/ui/components/Checklist.tsx
@@ -2,8 +2,15 @@
 
 import { Result } from "effect";
 import { useTranslations } from "next-intl";
-import { Card, Owner, ownerLabel } from "../../logic/GameObjects";
-import { allOwners, cardName } from "../../logic/GameSetup";
+import { useEffect, useState } from "react";
+import { Card, Owner, Player, ownerLabel } from "../../logic/GameObjects";
+import {
+    allCardIds,
+    allOwners,
+    cardName,
+    caseFileSize,
+    defaultHandSizes,
+} from "../../logic/GameSetup";
 import {
     Cell,
     CellValue,
@@ -33,29 +40,24 @@ import { Envelope } from "./Icons";
 import { Tooltip } from "./Tooltip";
 
 /**
- * The main visual: a case-file header strip on top; a grid with one row
- * per card and one column per owner underneath. Cells show Y / N / blank,
- * are coloured by status, and show a native browser tooltip (via the
- * `title` attribute) with the full explanation chain when you hover.
- * Blank cells that are still candidates for a refuter's unseen card
- * get footnote superscripts (the "number system"), also described via
- * the title tooltip.
+ * Unified tabbed checklist: the single surface for both editing the
+ * deck / roster (Setup mode) and tracking deductions (Play mode).
+ * State-slice ownership is one tab-gate deep: `inSetup` controls
+ * whether player name inputs, hand-size row, add / remove affordances,
+ * and the trailing "+" column render. The cell grid (Y / N / blank,
+ * tooltips, cross-highlighting, footnotes) is identical in both.
  *
- * We deliberately stick to `title` rather than a custom popover so the
- * hover affordance is the same everywhere — no difference between a
- * ✓ / · cell and a blank-with-footnote cell.
+ * The GameSetupPanel + ChecklistGrid pair this replaces is still
+ * mounted during commits 17–18 as a safety net and gets deleted in
+ * commit 19.
  */
 export function Checklist() {
     const t = useTranslations("deduce");
+    const tSetup = useTranslations("setup");
     const tReasons = useTranslations("reasons");
     const { state, dispatch, derived } = useClue();
     const { hoveredSuggestionIndex } = useHover();
-    // Tab mode gate — unused in commit 17 (this component is a literal
-    // copy of ChecklistGrid). Commit 18 folds the Setup-mode controls
-    // into the table rows and commit 19 deletes the ChecklistGrid /
-    // GameSetupPanel pair this replaces.
     const inSetup = state.uiMode === "setup";
-    void inSetup;
     const setup = state.setup;
     const knownCards = state.knownCards;
     const result = derived.deductionResult;
@@ -65,13 +67,34 @@ export function Checklist() {
 
     const owners: ReadonlyArray<Owner> = allOwners(setup);
 
+    const handSizeMap = new Map(state.handSizes);
+    const defaults = new Map(defaultHandSizes(setup));
+    const totalDealt = allCardIds(setup).length - caseFileSize(setup);
+    const setHandSizesArr = setup.players
+        .map(p => handSizeMap.get(p))
+        .filter((n): n is number => typeof n === "number");
+    const allHandSizesSet =
+        setHandSizesArr.length === setup.players.length &&
+        setup.players.length > 0;
+    const handSizesTotal = setHandSizesArr.reduce((a, b) => a + b, 0);
+    const handSizeMismatch =
+        allHandSizesSet && handSizesTotal !== totalDealt;
+
+    const onHandSizeChange = (player: Player, raw: string) => {
+        if (raw === "") {
+            dispatch({ type: "setHandSize", player, size: undefined });
+            return;
+        }
+        const n = Number(raw);
+        if (Number.isFinite(n) && n >= 0) {
+            dispatch({ type: "setHandSize", player, size: n });
+        }
+    };
+
     /**
      * Cross-highlight: when the user hovers a suggestion row in
      * PriorSuggestions, highlight every cell whose provenance chain
-     * referenced that suggestion's index. `chainFor` returns every
-     * Reason contributing to the cell's current value; any Reason
-     * whose `kind.suggestionIndex` matches the hovered index makes
-     * this cell participate.
+     * referenced that suggestion's index.
      */
     const cellIsHighlighted = (owner: Owner, card: Card): boolean => {
         if (hoveredSuggestionIndex === null) return false;
@@ -91,14 +114,9 @@ export function Checklist() {
     };
 
     /**
-     * Toggle a known-card entry for (player, card) when the user clicks a
-     * cell. Only player columns are interactive — the CaseFile column is
-     * computed by the deducer and never a direct user input.
-     *
-     * If the clicked (player, card) is already in knownCards, remove it;
-     * otherwise add it. If the cell currently shows N (deduced), clicking
-     * will add a Y known-card that contradicts — the global banner will
-     * show the user why.
+     * Toggle a known-card entry for (player, card) when the user clicks
+     * a cell. Only player columns are interactive — the case-file
+     * column is computed by the deducer.
      */
     const toggleKnownCard = (owner: Owner, card: Card) => {
         if (owner._tag !== "Player") return;
@@ -116,17 +134,12 @@ export function Checklist() {
         }
     };
 
-    // While the deducer is in a contradictory state, fall back to the
-    // empty-knowledge snapshot so the grid still renders (with the
-    // user's known-card inputs visible). The global contradiction banner
-    // at the top of the page surfaces the quick-fix UI; we don't block
-    // the grid anymore.
-    //
-    // We use Result.getOrUndefined rather than narrowing on isSuccess so
-    // React Compiler / Next Turbopack don't hoist a `.success` read ahead
-    // of the narrow check in their IR.
     const knowledge: Knowledge =
         Result.getOrUndefined(result) ?? emptyKnowledge;
+
+    // Column count for <th colSpan> on category / card-name / add-* rows.
+    // In Setup mode the trailing "+ add player" column adds one more.
+    const cardSpan = 1 + owners.length + (inSetup ? 1 : 0);
 
     return (
         <section className="min-w-0 rounded-[var(--radius)] border border-border bg-panel p-4">
@@ -134,6 +147,15 @@ export function Checklist() {
                 {t("title")}
             </h2>
             <CaseFileHeader knowledge={knowledge} />
+            {inSetup && handSizeMismatch && (
+                <div className="mb-3 rounded-[var(--radius)] border border-warning-border bg-warning-bg px-3 py-2 text-[13px] text-warning">
+                    {tSetup("handSizeMismatch", {
+                        total: handSizesTotal,
+                        expected: totalDealt,
+                        caseFileCount: caseFileSize(setup),
+                    })}
+                </div>
+            )}
             <table className="w-full border-collapse text-[13px]">
                 <thead>
                     <tr>
@@ -141,111 +163,303 @@ export function Checklist() {
                         {owners.map(owner => (
                             <th
                                 key={ownerKey(owner)}
-                                className="sticky top-0 z-10 border border-border bg-row-header px-2 py-1 text-center font-semibold"
+                                className="sticky top-0 z-10 border border-border bg-row-header px-2 py-1 text-center align-top font-semibold"
                             >
-                                {ownerLabel(owner)}
+                                {inSetup && owner._tag === "Player" ? (
+                                    <PlayerNameInput
+                                        player={owner.player}
+                                        allPlayers={setup.players}
+                                    />
+                                ) : (
+                                    ownerLabel(owner)
+                                )}
                             </th>
                         ))}
+                        {inSetup && (
+                            <th className="w-8 border border-border bg-row-header px-1.5 py-1 text-center">
+                                <button
+                                    type="button"
+                                    className="h-6 w-6 cursor-pointer rounded border-none bg-accent text-[16px] leading-none text-white hover:bg-accent-hover"
+                                    title={tSetup("addPlayerTitle")}
+                                    onClick={() =>
+                                        dispatch({ type: "addPlayer" })
+                                    }
+                                >
+                                    +
+                                </button>
+                            </th>
+                        )}
                     </tr>
+                    {inSetup && (
+                        <tr>
+                            <th className="whitespace-nowrap border border-border bg-row-header px-1.5 py-1 text-left font-semibold">
+                                {tSetup("handSize")}
+                            </th>
+                            {owners.map(owner => {
+                                if (owner._tag !== "Player") {
+                                    return (
+                                        <td
+                                            key={ownerKey(owner)}
+                                            className="border border-border"
+                                        />
+                                    );
+                                }
+                                const current = handSizeMap.get(owner.player);
+                                const def = defaults.get(owner.player);
+                                return (
+                                    <td
+                                        key={ownerKey(owner)}
+                                        className="border border-border px-1.5 py-1 text-center"
+                                    >
+                                        <input
+                                            type="number"
+                                            min={0}
+                                            max={allCardIds(setup).length}
+                                            className="w-14 rounded border border-border p-0.5 text-center text-[12px]"
+                                            value={
+                                                current === undefined
+                                                    ? ""
+                                                    : String(current)
+                                            }
+                                            placeholder={
+                                                def === undefined
+                                                    ? ""
+                                                    : String(def)
+                                            }
+                                            onChange={e =>
+                                                onHandSizeChange(
+                                                    owner.player,
+                                                    e.currentTarget.value,
+                                                )
+                                            }
+                                        />
+                                    </td>
+                                );
+                            })}
+                            <td className="border border-border" />
+                        </tr>
+                    )}
                 </thead>
                 <tbody>
-                    {setup.categories.flatMap(category => [
-                        <tr key={`h-${String(category.id)}`}>
-                            <th
-                                colSpan={1 + owners.length}
-                                className="border border-border bg-accent px-2 py-1.5 text-left text-[11px] uppercase tracking-[0.05em] text-white"
-                            >
-                                {category.name}
-                            </th>
-                        </tr>,
-                        ...category.cards.map(entry => (
-                            <tr key={String(entry.id)}>
-                                <th className="border border-border px-2 py-1 text-left font-normal">
-                                    {entry.name}
-                                </th>
-                                {owners.map(owner => {
-                                    const value = getCellByOwnerCard(
-                                        knowledge,
-                                        owner,
-                                        entry.id,
-                                    );
-                                    const footnoteNumbers = footnotesForCell(
-                                        footnotes,
-                                        Cell(owner, entry.id),
-                                    );
-                                    const isPlayerCell = owner._tag === "Player";
-                                    const isHighlighted = cellIsHighlighted(
-                                        owner,
-                                        entry.id,
-                                    );
-                                    const tooltipText = buildCellTitle({
-                                        provenance,
-                                        suggestions,
-                                        setup,
-                                        owner,
-                                        card: entry.id,
-                                        footnoteNumbers,
-                                        tDeduce: t,
-                                        tReasons,
-                                    });
-                                    const tooltipContent = tooltipText ? (
-                                        <div className="whitespace-pre-line">
-                                            {tooltipText}
-                                        </div>
-                                    ) : undefined;
-                                    return (
-                                        <Tooltip
-                                            key={`${ownerKey(owner)}-${String(entry.id)}`}
-                                            content={tooltipContent}
-                                        >
-                                            <td
-                                                className={cellClass(
-                                                    value,
-                                                    isPlayerCell,
-                                                    isHighlighted,
-                                                )}
-                                                onClick={
-                                                    isPlayerCell
-                                                        ? () =>
-                                                              toggleKnownCard(
-                                                                  owner,
-                                                                  entry.id,
-                                                              )
-                                                        : undefined
+                    {setup.categories.flatMap(category => {
+                        const canRemoveCategory = setup.categories.length > 1;
+                        const canRemoveCard = category.cards.length > 1;
+                        return [
+                            <tr key={`h-${String(category.id)}`}>
+                                <th
+                                    colSpan={cardSpan}
+                                    className="border border-border bg-accent px-2 py-1.5 text-left text-[11px] uppercase tracking-[0.05em] text-white"
+                                >
+                                    {inSetup ? (
+                                        <div className="flex items-center justify-between gap-2">
+                                            <InlineTextEdit
+                                                value={category.name}
+                                                className="min-w-0 flex-1 rounded border border-white/30 bg-transparent px-1 py-0.5 text-[11px] font-semibold uppercase tracking-[0.05em] text-white focus:bg-white/10 focus:outline-none"
+                                                title={tSetup("renameCategoryTitle")}
+                                                onCommit={next =>
+                                                    dispatch({
+                                                        type: "renameCategory",
+                                                        categoryId: category.id,
+                                                        name: next,
+                                                    })
                                                 }
-                                                role={isPlayerCell ? "button" : undefined}
-                                                tabIndex={isPlayerCell ? 0 : undefined}
-                                                onKeyDown={
-                                                    isPlayerCell
-                                                        ? e => {
-                                                              if (
-                                                                  e.key === "Enter" ||
-                                                                  e.key === " "
-                                                              ) {
-                                                                  e.preventDefault();
+                                            />
+                                            <button
+                                                type="button"
+                                                title={
+                                                    canRemoveCategory
+                                                        ? tSetup("removeCategoryTitle", {
+                                                              name: category.name,
+                                                          })
+                                                        : tSetup("removeCategoryMin")
+                                                }
+                                                disabled={!canRemoveCategory}
+                                                className="cursor-pointer border-none bg-transparent p-0 text-[14px] leading-none text-white/80 hover:text-white disabled:cursor-not-allowed disabled:opacity-40"
+                                                onClick={() =>
+                                                    dispatch({
+                                                        type: "removeCategoryById",
+                                                        categoryId: category.id,
+                                                    })
+                                                }
+                                            >
+                                                &times;
+                                            </button>
+                                        </div>
+                                    ) : (
+                                        category.name
+                                    )}
+                                </th>
+                            </tr>,
+                            ...category.cards.map(entry => (
+                                <tr key={String(entry.id)}>
+                                    <th className="border border-border px-2 py-1 text-left font-normal">
+                                        {inSetup ? (
+                                            <div className="flex items-center justify-between gap-2">
+                                                <InlineTextEdit
+                                                    value={entry.name}
+                                                    className="min-w-0 flex-1 rounded border border-transparent bg-transparent px-1 py-0.5 text-[12px] hover:border-border focus:border-accent focus:outline-none"
+                                                    title={tSetup("renameCardTitle")}
+                                                    onCommit={next =>
+                                                        dispatch({
+                                                            type: "renameCard",
+                                                            cardId: entry.id,
+                                                            name: next,
+                                                        })
+                                                    }
+                                                />
+                                                <button
+                                                    type="button"
+                                                    title={
+                                                        canRemoveCard
+                                                            ? tSetup("removeCardTitle", {
+                                                                  name: entry.name,
+                                                              })
+                                                            : tSetup("removeCardMin")
+                                                    }
+                                                    disabled={!canRemoveCard}
+                                                    className="cursor-pointer border-none bg-transparent p-0 text-[14px] leading-none text-muted hover:text-danger disabled:cursor-not-allowed disabled:opacity-40"
+                                                    onClick={() =>
+                                                        dispatch({
+                                                            type: "removeCardById",
+                                                            cardId: entry.id,
+                                                        })
+                                                    }
+                                                >
+                                                    &times;
+                                                </button>
+                                            </div>
+                                        ) : (
+                                            entry.name
+                                        )}
+                                    </th>
+                                    {owners.map(owner => {
+                                        const value = getCellByOwnerCard(
+                                            knowledge,
+                                            owner,
+                                            entry.id,
+                                        );
+                                        const footnoteNumbers = footnotesForCell(
+                                            footnotes,
+                                            Cell(owner, entry.id),
+                                        );
+                                        const isPlayerCell = owner._tag === "Player";
+                                        const isHighlighted = cellIsHighlighted(
+                                            owner,
+                                            entry.id,
+                                        );
+                                        const tooltipText = buildCellTitle({
+                                            provenance,
+                                            suggestions,
+                                            setup,
+                                            owner,
+                                            card: entry.id,
+                                            footnoteNumbers,
+                                            tDeduce: t,
+                                            tReasons,
+                                        });
+                                        const tooltipContent = tooltipText ? (
+                                            <div className="whitespace-pre-line">
+                                                {tooltipText}
+                                            </div>
+                                        ) : undefined;
+                                        return (
+                                            <Tooltip
+                                                key={`${ownerKey(owner)}-${String(entry.id)}`}
+                                                content={tooltipContent}
+                                            >
+                                                <td
+                                                    className={cellClass(
+                                                        value,
+                                                        isPlayerCell,
+                                                        isHighlighted,
+                                                    )}
+                                                    onClick={
+                                                        isPlayerCell
+                                                            ? () =>
                                                                   toggleKnownCard(
                                                                       owner,
                                                                       entry.id,
-                                                                  );
+                                                                  )
+                                                            : undefined
+                                                    }
+                                                    role={isPlayerCell ? "button" : undefined}
+                                                    tabIndex={isPlayerCell ? 0 : undefined}
+                                                    onKeyDown={
+                                                        isPlayerCell
+                                                            ? e => {
+                                                                  if (
+                                                                      e.key === "Enter" ||
+                                                                      e.key === " "
+                                                                  ) {
+                                                                      e.preventDefault();
+                                                                      toggleKnownCard(
+                                                                          owner,
+                                                                          entry.id,
+                                                                      );
+                                                                  }
                                                               }
-                                                          }
-                                                        : undefined
-                                                }
-                                            >
-                                                {cellLabel(value)}
-                                                {footnoteNumbers.length > 0 &&
-                                                    value === undefined && (
-                                                        <sup className="ml-0.5 text-[9px] font-normal text-accent">
-                                                            {footnoteNumbers.join(",")}
-                                                        </sup>
-                                                    )}
-                                            </td>
-                                        </Tooltip>
-                                    );
-                                })}
-                            </tr>
-                        )),
-                    ])}
+                                                            : undefined
+                                                    }
+                                                >
+                                                    {cellLabel(value)}
+                                                    {footnoteNumbers.length > 0 &&
+                                                        value === undefined && (
+                                                            <sup className="ml-0.5 text-[9px] font-normal text-accent">
+                                                                {footnoteNumbers.join(",")}
+                                                            </sup>
+                                                        )}
+                                                </td>
+                                            </Tooltip>
+                                        );
+                                    })}
+                                    {inSetup && (
+                                        <td className="border border-border" />
+                                    )}
+                                </tr>
+                            )),
+                            ...(inSetup
+                                ? [
+                                      <tr key={`add-card-${String(category.id)}`}>
+                                          <th
+                                              colSpan={cardSpan}
+                                              className="border border-border bg-row-alt px-1.5 py-1 text-left"
+                                          >
+                                              <button
+                                                  type="button"
+                                                  className="cursor-pointer border-none bg-transparent p-0 text-[12px] text-accent underline"
+                                                  onClick={() =>
+                                                      dispatch({
+                                                          type: "addCardToCategoryById",
+                                                          categoryId: category.id,
+                                                      })
+                                                  }
+                                              >
+                                                  {tSetup("addCard")}
+                                              </button>
+                                          </th>
+                                      </tr>,
+                                  ]
+                                : []),
+                        ];
+                    })}
+                    {inSetup && (
+                        <tr>
+                            <th
+                                colSpan={cardSpan}
+                                className="border border-border bg-row-alt px-1.5 py-2 text-center"
+                            >
+                                <button
+                                    type="button"
+                                    className="cursor-pointer rounded border border-border bg-white px-3 py-1 text-[13px] hover:bg-hover"
+                                    onClick={() =>
+                                        dispatch({ type: "addCategory" })
+                                    }
+                                >
+                                    {tSetup("addCategory")}
+                                </button>
+                            </th>
+                        </tr>
+                    )}
                 </tbody>
             </table>
         </section>
@@ -253,10 +467,133 @@ export function Checklist() {
 }
 
 /**
+ * Editable text cell. Commits the new value on blur or Enter; resets
+ * to the external value on Escape or if the input is cleared.
+ */
+function InlineTextEdit({
+    value,
+    onCommit,
+    className,
+    title,
+}: {
+    value: string;
+    onCommit: (next: string) => void;
+    className?: string;
+    title?: string;
+}) {
+    const [local, setLocal] = useState(value);
+    useEffect(() => {
+        setLocal(value);
+    }, [value]);
+
+    const commit = () => {
+        const trimmed = local.trim();
+        if (trimmed.length === 0) {
+            setLocal(value);
+            return;
+        }
+        if (trimmed !== value) onCommit(trimmed);
+    };
+
+    return (
+        <input
+            type="text"
+            value={local}
+            className={className}
+            title={title}
+            onChange={e => setLocal(e.currentTarget.value)}
+            onBlur={commit}
+            onKeyDown={e => {
+                if (e.key === "Enter") {
+                    (e.currentTarget as HTMLInputElement).blur();
+                } else if (e.key === "Escape") {
+                    setLocal(value);
+                    (e.currentTarget as HTMLInputElement).blur();
+                }
+            }}
+        />
+    );
+}
+
+/**
+ * Editable player-name header with remove-× button. Handles the
+ * duplicate-name check locally so the reducer doesn't have to.
+ */
+function PlayerNameInput({
+    player,
+    allPlayers,
+}: {
+    player: Player;
+    allPlayers: ReadonlyArray<Player>;
+}) {
+    const t = useTranslations("setup");
+    const { dispatch } = useClue();
+    const [editing, setEditing] = useState(String(player));
+    const [error, setError] = useState("");
+
+    useEffect(() => {
+        setEditing(String(player));
+        setError("");
+    }, [player]);
+
+    const commit = () => {
+        const trimmed = editing.trim();
+        if (!trimmed) {
+            setEditing(String(player));
+            setError("");
+            return;
+        }
+        if (trimmed === String(player)) {
+            setError("");
+            return;
+        }
+        if (allPlayers.some(p => String(p) === trimmed)) {
+            setError(t("duplicateName"));
+            return;
+        }
+        dispatch({
+            type: "renamePlayer",
+            oldName: player,
+            newName: Player(trimmed),
+        });
+        setError("");
+    };
+
+    return (
+        <div className="flex flex-col items-stretch gap-0.5">
+            <input
+                type="text"
+                className="box-border w-full rounded border border-border px-1.5 py-1 text-[12px]"
+                value={editing}
+                onChange={e => {
+                    setEditing(e.currentTarget.value);
+                    setError("");
+                }}
+                onBlur={commit}
+                onKeyDown={e => {
+                    if (e.key === "Enter") commit();
+                }}
+            />
+            {error && (
+                <span className="whitespace-nowrap text-[11px] text-danger">
+                    {error}
+                </span>
+            )}
+            <button
+                type="button"
+                className="self-center border-none bg-transparent px-1 text-[14px] leading-none text-muted hover:text-danger"
+                title={t("removePlayerTitle", { player: String(player) })}
+                onClick={() => dispatch({ type: "removePlayer", player })}
+            >
+                &times;
+            </button>
+        </div>
+    );
+}
+
+/**
  * Resolve a single `ReasonDescription` (from `describeReason`) into
  * `{ headline, detail }` strings via the "reasons" i18n namespace.
- * Centralising the lookup here keeps the cell-title builder compact
- * and makes the shape of each reason variant visible in one place.
  */
 const resolveReasonCopy = (
     desc: ReasonDescription,
@@ -337,13 +674,6 @@ const resolveReasonCopy = (
     }
 };
 
-/**
- * Assemble the title= string shown on hover. For known Y/N cells we walk
- * the dependency chain backwards and render each step as a numbered line
- * so the user sees *why* the cell has that value, not just the last
- * rule. For blank cells with refuter-candidate footnotes we explain the
- * footnote numbers. For everything else, no tooltip at all.
- */
 const buildCellTitle = (args: {
     provenance: Provenance | undefined;
     suggestions: ReadonlyArray<Suggestion>;
@@ -476,9 +806,6 @@ const CELL_BASE =
 const CELL_INTERACTIVE =
     " cursor-pointer hover:ring-2 hover:ring-accent/40 focus:outline-none focus:ring-2 focus:ring-accent";
 
-// Persistent cross-highlight (from hovering a suggestion in the log).
-// Stronger ring + subtle offset so it visually distinguishes from the
-// hover/focus ring on interactive cells and survives both.
 const CELL_HIGHLIGHTED =
     " ring-2 ring-accent ring-offset-1 ring-offset-panel";
